+++ feature/onboarding/src/main/java/com/twix/onboarding/di/OnBoardingModule.kt
@@ -2,8 +2,8 @@ package com.twix.onboarding.di
 
 import com.twix.navigation.NavRoutes
 import com.twix.navigation.base.NavGraphContributor
+import com.twix.onboarding.OnBoardingViewModel
 import com.twix.onboarding.navigation.OnboardingNavGraph
-import com.twix.onboarding.vm.OnBoardingViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
