+++ feature/onboarding/src/main/java/com/twix/onboarding/invite/InviteCodeScreen.kt
@@ -47,11 +47,11 @@ import com.twix.designsystem.theme.CommonColor
 import com.twix.designsystem.theme.GrayColor
 import com.twix.designsystem.theme.TwixTheme
 import com.twix.domain.model.enums.AppTextStyle
+import com.twix.onboarding.OnBoardingViewModel
 import com.twix.onboarding.R
 import com.twix.onboarding.invite.component.InviteCodeTextField
 import com.twix.onboarding.model.OnBoardingIntent
 import com.twix.onboarding.model.OnBoardingSideEffect
-import com.twix.onboarding.vm.OnBoardingViewModel
 import com.twix.ui.base.ObserveAsEvents
 import com.twix.ui.extension.noRippleClickable
 import com.twix.ui.keyboard.Keyboard
